Move retrieve_project_github function to test utils
