+++ asreview/webapp/tests/test_project.py
@@ -17,7 +17,7 @@ import pytest
 from io import BytesIO
 from urllib.request import urlopen
 
-from asreview.webapp.utils.project import retrieve_project_url_github
+from asreview.webapp.tests.utils import retrieve_project_url_github
 
 
 # Retrieve urls to .asreview files exported from previous versions
